added file extension .ts to require
