+++ packages/zipkin/test/jsonEncoder.JSON_V1.test.ts
@@ -1,18 +1,19 @@
-const TraceId = require('../src/tracer/TraceId');
-const {Span, Endpoint} = require('../src/model');
-const {JSON_V2} = require('../src/jsonEncoder');
-const {Some, None} = require('../src/option');
+const TraceId = require('../src/tracer/TraceId.ts');
+const {Span, Endpoint} = require('../src/model.ts');
+const {JSON_V1} = require('../src/jsonEncoder.ts');
+const {Some, None} = require('../src/option.ts');
 
-describe('JSON v2 Formatting', () => {
-  it('should write minimum fields', () => {
+describe('JSON v1 Formatting', () => {
+  // v1 format requires an empty span name. v2 can leave it out
+  it('should write minimum fields, notably an empty span name', () => {
     const span = new Span(new TraceId({
       traceId: new Some('000000000000162e'),
       spanId: '000000000000162e',
       sampled: None
     }));
 
-    expect(JSON_V2.encode(span)).to.equal(
-      '{"traceId":"000000000000162e","id":"000000000000162e"}'
+    expect(JSON_V1.encode(span)).to.equal(
+      '{"traceId":"000000000000162e","id":"000000000000162e","name":""}'
     );
   });
 
@@ -24,7 +25,7 @@ describe('JSON v2 Formatting', () => {
       sampled: None
     }));
 
-    expect(JSON_V2.encode(span)).to.contain(
+    expect(JSON_V1.encode(span)).to.contain(
       '{"traceId":"000000000000162e","parentId":"000000000000abcd"'
     );
   });
@@ -35,7 +36,7 @@ describe('JSON v2 Formatting', () => {
       spanId: '000000000000162e'
     }));
 
-    expect(JSON_V2.encode(span)).to.contain(
+    expect(JSON_V1.encode(span)).to.contain(
       '{"traceId":"00000000000004d2000000000000162e"'
     );
   });
@@ -47,7 +48,7 @@ describe('JSON v2 Formatting', () => {
       flags: 1
     }));
 
-    expect(JSON_V2.encode(span)).to.contain(',"debug":true}');
+    expect(JSON_V1.encode(span)).to.contain(',"debug":true}');
   });
 
   it('should transform correctly from Span to JSON representation', () => {
@@ -67,46 +68,72 @@ describe('JSON v2 Formatting', () => {
     span.setTimestamp(1);
     span.setDuration(1);
     span.putTag('warning', 'The cake is a lie');
+    span.setShared(true);
 
     const expected = {
       traceId: 'a',
       name: 'get',
       id: 'c',
       parentId: 'b',
-      kind: 'SERVER',
-      timestamp: 1,
-      duration: 1,
-      localEndpoint: {
-        serviceName: 'portalservice',
-        ipv4: '10.57.50.83',
-        port: 8080
-      },
-      tags: {warning: 'The cake is a lie'}
+      annotations: [
+        {
+          endpoint: {
+            serviceName: 'portalservice',
+            ipv4: '10.57.50.83',
+            port: 8080
+          },
+          timestamp: 1,
+          value: 'sr'
+        },
+        {
+          endpoint: {
+            serviceName: 'portalservice',
+            ipv4: '10.57.50.83',
+            port: 8080,
+          },
+          timestamp: 2,
+          value: 'ss'
+        }
+      ],
+      binaryAnnotations: [
+        {
+          key: 'warning',
+          value: 'The cake is a lie',
+          endpoint: {
+            serviceName: 'portalservice',
+            ipv4: '10.57.50.83',
+            port: 8080
+          }
+        }
+      ]
     };
 
-    const spanJson = JSON.parse(JSON_V2.encode(span));
+    const spanJson = JSON.parse(JSON_V1.encode(span));
     expect(spanJson.traceId).to.equal(expected.traceId);
-    expect(spanJson.parentId).to.equal(expected.parentId);
-    expect(spanJson.id).to.equal(expected.id);
     expect(spanJson.name).to.equal(expected.name);
-    expect(spanJson.kind).to.equal(expected.kind);
-    expect(spanJson.timestamp).to.equal(expected.timestamp);
-    expect(spanJson.duration).to.equal(expected.duration);
-    expect(spanJson.tags).to.deep.equal(expected.tags);
+    expect(spanJson.id).to.equal(expected.id);
+    expect(spanJson.parentId).to.equal(expected.parentId);
+    expect(spanJson.annotations).to.deep.equal(expected.annotations);
+    expect(spanJson.binaryAnnotations).to.deep.equal(expected.binaryAnnotations);
   });
 
-  it('should set shared', () => {
+  it('should not set timestamp or duration on shared span', () => {
     const span = new Span(new TraceId({
       traceId: new Some('a'),
       parentId: new Some('b'),
       spanId: 'c',
       sampled: None
     }));
+    span.setName('GET');
+    span.setKind('SERVER');
+    span.setTimestamp(1);
+    span.setDuration(1);
     span.setShared(true);
 
-    const spanJson = JSON_V2.encode(span);
+    const spanJson = JSON_V1.encode(span);
 
-    expect(spanJson).to.contain('"shared":true');
+    expect(spanJson).to.not.contain('"name":"GET","timestamp"');
+    expect(spanJson).to.not.contain('"duration"');
   });
 
   it('should set timestamp and duration on client span', () => {
@@ -121,10 +148,10 @@ describe('JSON v2 Formatting', () => {
     span.setTimestamp(1);
     span.setDuration(1);
 
-    expect(JSON_V2.encode(span)).to.contain('"timestamp":1,"duration":1}');
+    expect(JSON_V1.encode(span)).to.contain('"timestamp":1,"duration":1,');
   });
 
-  it('should set remoteEndpoint on client span', () => {
+  it('should set server address on client span', () => {
     const span = new Span(new TraceId({
       traceId: new Some('a'),
       parentId: new Some('b'),
@@ -139,14 +166,14 @@ describe('JSON v2 Formatting', () => {
       port: 80
     }));
 
-    const spanJson = JSON_V2.encode(span);
+    const spanJson = JSON_V1.encode(span);
     expect(spanJson).to.contain(
-      '"remoteEndpoint":{"serviceName":"there","ipv4":"10.57.50.84","port":80}}'
+      '{"key":"sa","value":true,"endpoint":{"serviceName":"there","ipv4":"10.57.50.84","port":80}}'
     );
   });
 
   // make sure nothing strange happens like object interpretation of dots
-  it('should serialize tags with dotted names', () => {
+  it('should serialize tags with dotted names as binary annotations', () => {
     const span = new Span(new TraceId({
       traceId: new Some('a'),
       parentId: new Some('b'),
@@ -156,7 +183,9 @@ describe('JSON v2 Formatting', () => {
     span.setName('GET');
     span.putTag('http.path', '/api');
 
-    const spanJson = JSON_V2.encode(span);
-    expect(spanJson).to.contain('"tags":{"http.path":"/api"}}');
+    const spanJson = JSON_V1.encode(span);
+    expect(spanJson).to.contain(
+      '"binaryAnnotations":[{"key":"http.path","value":"/api"}]}'
+    );
   });
 });
