+++ packages/zipkin/test/trace.test.ts
@@ -2,12 +2,12 @@ const sinon = require('sinon');
 const lolex = require('lolex');
 const Promise = require('bluebird');
 const isPromise = require('is-promise');
-const Tracer = require('../src/tracer');
-const Annotation = require('../src/annotation');
-const {Sampler, neverSample} = require('../src/tracer/sampler');
-const ExplicitContext = require('../src/explicit-context');
-const {Some} = require('../src/option');
-const {Endpoint} = require('../src/model');
+const Tracer = require('../src/tracer/index.ts');
+const Annotation = require('../src/annotation.ts');
+const {Sampler, neverSample} = require('../src/tracer/sampler.ts');
+const ExplicitContext = require('../src/explicit-context.ts');
+const {Some} = require('../src/option.ts');
+const {Endpoint} = require('../src/model.ts');
 
 describe('Tracer', () => {
   it('should make parent and child spans', () => {
